+++ iOS/Developer/Frameworks/XCTest.framework/XCUIAXNotificationHandling-Protocol.h
@@ -10,6 +10,6 @@
 
 @protocol XCUIAXNotificationHandling <NSObject>
 @property(readonly) _Bool axNotificationsIncludeElement;
-- (void)handleAccessibilityNotification:(long long)arg1 fromElement:(XCAccessibilityElement *)arg2 payload:(NSData *)arg3;
+- (void)handleAccessibilityNotification:(int)arg1 fromElement:(XCAccessibilityElement *)arg2 payload:(NSData *)arg3;
 @end
