+++ iOS/Frameworks/PhotosUI.framework/PUPhotosPanoramaViewControllerSpec.h
@@ -14,8 +14,8 @@
 - (id)gridSpec;
 - (void)configureCollectionViewGridLayout:(id)arg1 referenceWidth:(double)arg2 safeAreaInsets:(struct UIEdgeInsets)arg3;
 - (long long)cellFillMode;
-- (int)fastThumbnailImageFormat;
-- (int)thumbnailImageFormat;
+- (unsigned short)fastThumbnailImageFormat;
+- (unsigned short)thumbnailImageFormat;
 - (double)globalFooterHeight;
 - (double)sectionHeaderHeight;
 - (unsigned long long)supportedInterfaceOrientations;
