+++ iOS/PrivateFrameworks/iCalendar.framework/ICSFreeBusyTime.h
@@ -8,7 +8,7 @@
 {
 }
 
-@property(nonatomic) int fbtype;
+@property(nonatomic) unsigned long long fbtype;
 - (id)initWithPeriod:(id)arg1;
 
 @end
