+++ iOS/PrivateFrameworks/iWorkImport.framework/TSWPStyleDiff.h
@@ -25,7 +25,7 @@ __attribute__((visibility("hidden")))
 - (_Bool)isEqual:(id)arg1;
 - (_Bool)isEqualToStyleDiff:(id)arg1;
 - (id)changedPropertyNames:(id)arg1;
-- (id)styleDiffForTable:(int)arg1 storage:(id)arg2 currentObject:(id)arg3;
+- (id)styleDiffForTable:(unsigned long long)arg1 storage:(id)arg2 currentObject:(id)arg3;
 - (void)saveToArchiver:(id)arg1;
 - (void)saveToArchive:(struct StyleDiffArchive *)arg1 archiver:(id)arg2;
 - (void)loadFromUnarchiver:(id)arg1;
